+++ source/BuildingBlocks.Infrastructure/BuildingBlocks.Infrastructure.csproj
@@ -5,5 +5,6 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Data.SqlClient" Version="5.1.2" />
+    <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.16.1" />
   </ItemGroup>
 </Project>
