+++ source/IntegrationTests/TestDoubles/RemoteBusinessServiceRequestSenderSpy.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Threading.Tasks;
 using Azure.Messaging.ServiceBus;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.MessageBus.RemoteBusinessServices;
+using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.MessageBus.RemoteBusinessServices;
 
 namespace Energinet.DataHub.EDI.IntegrationTests.TestDoubles
 {
